Comleting Task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,15 +11,15 @@
  */
 int main(void)
 {
-	int i, n1 = 1, n2 = 2, sum;
+	long int i, n1 = 1, n2 = 2, sum;
 
-	printf("%d, %d", n1, n2);
+	printf("%lu, %d", n1, n2);
 	for (i = 1; i <= 50; i++)
 	{
 		sum = n1 + n2;
 		n1 = n2;
 		n2 = sum;
-		printf(", %d", sum);
+		printf(", %lu", sum);
 	}
 	printf("\n");
 	return (0);
